% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afdx.R
\docType{package}
\name{afdx-package}
\alias{afdx}
\alias{afdx-package}
\title{afdx: Diagnosis performance indicators from attributable fraction estimates.}
\description{
The afdx package provides functions to estimate the logit exponential model
and the bayesian latent class model.
}
\section{The logit exponential model}{


The logitexp function estimated the logit exponential function fitting a
maximum likelihood model. The senspec functions estimate the sensitivity,
specificity, positive predicted value and negative predicted values for
the specified cut-off points.
}

\section{The bayesian latent class model}{


The get_latent_model provides an rjags function template to estimate
the attributable fraction and the sensitivity, specificity, positive
predicted value and negative predicted value of the latent class model.

@docType package
@name afdx
}

\author{
\strong{Maintainer}: John J. Aponte \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})

Authors:
\itemize{
  \item Orvalho Augusto \email{caveman@gmail.com} (\href{https://orcid.org/0000-0002-0005-3968}{ORCID})
}

}
