% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latent_model.R
\name{get_latent_model}
\alias{get_latent_model}
\title{Template for the bayesian latent class model}
\usage{
get_latent_model()
}
\value{
a string value
}
\description{
This function returns a template that can be use as model in an rJags model
it requires two vectors with the number of subjects in the symptoms,
like fever in the case of malaria (n) and the number of non-symptomatic (m)
in each of the categories of results of the diagnostic test. The first
category is reserved for the negatives by the diagnostic test (in the malaria
case those with asexual density 0) and the rest categories each one with
higher values than the previous category.
}
\details{
See:
Smith T, Vounatsou P. Logistic regression and latent class models for
estimating positivities in diagnostic assays with poor resolution.
Communications in Statistics - Theory and Methods. 1997 Jan;26(7):1677–700.

Vounatsou P, Smith T, Smith AFM. Bayesian analysis of two-component mixture
distributions applied to estimating malaria attributable fractions. Journal
of the Royal Statistical Society: Series C (Applied Statistics).
1998;47(4):575–87.

Müller I, Genton B, Rare L, Kiniboro B, Kastens W, Zimmerman P, et al.
Three different Plasmodium species show similar patterns of clinical
tolerance of malaria infection. Malar J. 2009;8(1):158.

Plucinski MM, Rogier E, Dimbu PR, Fortes F, Halsey ES, Aidoo M, et al.
Performance of Antigen Concentration Thresholds for Attributing Fever to
Malaria among Outpatients in Angola. J Clin Microbiol. 2019;57(3).
}
