% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cutoffs.R
\name{make_cutoffs}
\alias{make_cutoffs}
\title{Cut-off points for densities and fever}
\usage{
make_cutoffs(v.fever, v.density, add1 = TRUE)
}
\arguments{
\item{v.fever}{numeric vector of 0/1 indicating fever or equivalent}

\item{v.density}{numeric vector of values >= 0 indicating the density}
}
\description{
Generate the cutoffs at every change of density in the fever, but
first category is for density 0, and last category if possible have
no subjects with no fever.
}
