% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cutoffs.R
\name{make_n_cutoffs}
\alias{make_n_cutoffs}
\title{Make a defined number of categories having similar number of positives in each category}
\usage{
make_n_cutoffs(v.fever, v.density, mintot, add1 = TRUE)
}
\arguments{
\item{v.density}{numeric vector of values >= 0 indicating the density}

\item{ncat}{number of category}
}
\value{
a vector with the proposed cutoff points
}
\description{
Generate the a define number of categories with similar number of positive
in each category
}
