% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malaria_df2}
\alias{malaria_df2}
\title{Synthetic data simulating a malaria crossectional}
\format{
{ a dataset with two variables
\describe{
\item{fever}{1 if fever or history of fever, 0 otherwise}
\item{density}{asexual plasmodium parasite density, in parasites per ul}
}
}
}
\usage{
malaria_df2
}
\description{
Simulated data with the main outcomes of a malaria crossectional,
fever and parasite density
}
\keyword{datasets}
